// types/note.ts
export interface FetchNotesResponse {
	notes: Note[];
	totalPages: number;
}

export interface Note {
	id: string;
	title: string;
	content: string;
	tag: NoteTag;
	createdAt: string;
	updatedAt: string;
}

export type NoteTag = "Todo" | "Work" | "Personal" | "Meeting" | "Shopping";

export interface FormValues {
	title: string;
	content: string;
	tag: NoteTag;
}